updates authentication + authorization logic
that is, upon succesful authentication respond by with some authorization data.

We're abusing jwt to structure this and to manage the session.  It's Hideously Functional⁽™⁾.

Basically, a registered user validates their user/pass at the /auth/login endpoint using HTTP
Basic Auth.  They then receive a jwt-token for subsequent use when making requests to protected endpoints.

As a proof-of-concept, the following endpoints should be protected:

/info       -- accessible with HTTP Basic Auth (only)
/auth/scope -- accessible via JWT Bearer token (only)