update validation schemas;
the spec / interface for GET and POST differ slightly in that query
structure for the POST endpoint also has properties defined by the
framework, and not just the endpoint.

that is,
 paramaters for GET are derived from:
   .../schema/openapi/ga4gh-beacon/beacon-v2/models/json/beacon-v2-default-model/genomicVariations/endpoints.json

 and those for POST are derived as a combination of the following:
   .../schema/openapi/ga4gh-beacon/beacon-v2/models/json/beacon-v2-default-model/genomicVariations/endpoints.json
   .../schema/openapi/ga4gh-beacon/beacon-v2/framework/json/requests/beaconRequestBody.json

We've also cheated a little bit and used Joi.ref() to save code
duplication.  But MIGHT need a specification review... SHOULD be ok.
(abusing the terms).