updates apiClient wrapper;
although probably not worth it; was hoping to use axios.interceptors to
have a global default and then override. which does work, but it doesn't
really reduce the overall complexity at the point of usage.

i.e., axios.call().then().catch() still needed. Probs going to switch to
globally sources response and error parsing to encapsulate / specify the "client-api".